#### PR description:

As discussed here: https://indico.cern.ch/event/926508/#9-review-of-pf-layer-clusterin
we often effectively create PFClusters directly from PFRecHit seed passing the seeds threshold for HF (e.g. at HLT). However, in the current code and configuration setup, we still go through the topo clustering step and iterative clustering step without doing actual clustering, and wasting sizable time.

With this new cluster producer from individual seeds, we get basically identical physics outputs (except tiny mismatches due to limited numerical precision in the current setup). And, this step becomes faster by a factor of 6-7 (check HF line below).

Using Run2018D/EphemeralHLTPhysics data:

Before this PR
```
FastReport   CPU time avg.      when run  Real time avg.      when run     Alloc. avg.      when run   Dealloc. avg.      when run  Modules
FastReport         1.3 ms         1.3 ms         1.3 ms         1.3 ms        +710 kB        +710 kB        -599 kB        -599 kB    hltParticleFlowClusterECALUncorrectedUnseeded
FastReport         5.6 ms         5.6 ms         5.6 ms         5.6 ms       +3659 kB       +3659 kB       -2886 kB       -2886 kB    hltParticleFlowClusterPSUnseeded
FastReport         4.9 ms         4.9 ms         4.9 ms         4.9 ms        +122 kB        +122 kB         -30 kB         -30 kB    hltParticleFlowClusterECALUnseeded
FastReport         3.0 ms         3.0 ms         3.0 ms         3.0 ms        +671 kB        +671 kB        -587 kB        -587 kB    hltParticleFlowClusterHBHE
FastReport         0.4 ms         0.4 ms         0.4 ms         0.4 ms        +274 kB        +274 kB        -197 kB        -197 kB    hltParticleFlowClusterHCAL
FastReport         5.2 ms         5.2 ms         5.2 ms         5.2 ms       +1670 kB       +1670 kB       -1247 kB       -1247 kB    hltParticleFlowClusterHF
```

With this PR
```
FastReport         1.3 ms         1.3 ms         1.3 ms         1.3 ms        +710 kB        +710 kB        -599 kB        -599 kB    hltParticleFlowClusterECALUncorrectedUnseeded
FastReport         5.6 ms         5.6 ms         5.6 ms         5.6 ms       +3659 kB       +3659 kB       -2886 kB       -2886 kB    hltParticleFlowClusterPSUnseeded
FastReport         4.9 ms         4.9 ms         4.9 ms         4.9 ms        +122 kB        +122 kB         -30 kB         -30 kB    hltParticleFlowClusterECALUnseeded
FastReport         3.0 ms         3.0 ms         3.0 ms         3.0 ms        +671 kB        +671 kB        -587 kB        -587 kB    hltParticleFlowClusterHBHE
FastReport         0.4 ms         0.4 ms         0.4 ms         0.4 ms        +274 kB        +274 kB        -197 kB        -197 kB    hltParticleFlowClusterHCAL
FastReport         0.7 ms         0.7 ms         0.7 ms         0.7 ms       +1154 kB       +1154 kB        -886 kB        -886 kB    hltParticleFlowClusterHF
```

#### PR validation:

Checked with with Run2018D/EphemeralHLTPhysics data 100 events that, HLT PF cluster and PF candidates stay the same within numerical precision.
Check with matrix tests: 10824.0, 11634.0, 23234.0, and 140.53 that they run fine.

#### if this PR is a backport please specify the original PR and why you need to backport that PR:

This is not backport.

@bendavid @jsalfeld @dimkarasav