@cmsbuild 2020-08-11
A new Pull Request was created by  @hatakeyamak (Kenichi Hatakeyama) for master.

It involves the following packages:

HLTrigger/Configuration
RecoParticleFlow/PFClusterProducer


@perrotta, @Martin-Grunewald, @slava77, @cmsbuild, @fwyzard, @jpata can you please review it and eventually sign? Thanks.
@mmarionncern, @makortel, @felicepantaleo, @riga, @rovere, @lgray, @Martin-Grunewald, @clelange, @lecriste, @cbernet, @seemasharmafnal this is something you requested to watch as well.
@silviodonato, @dpiparo, @qliphy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>