@hatakeyamak 2020-08-11
> Hi,
> You only consider `hltParticleFlowClusterHF` but there are several more instances of `"PFClusterProducer"` within HLT menus:
> 
> ```
> cat OnLine_HLT_FULL.py | grep '"PFClusterProducer"'
> process.hltParticleFlowClusterECALUncorrectedUnseeded = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterPSUnseeded = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterHBHE = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterHF = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterPSL1Seeded = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterECALUncorrectedL1Seeded = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterECALUncorrectedForMuonsMF = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterPSForMuons = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterHFForEgammaUnseeded = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterECALUncorrectedForMuonsMFNoVtx = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterPSForMuonsNoVtx = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterECALUncorrectedForTkMuonsMF = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterPSForTkMuons = cms.EDProducer( "PFClusterProducer,
> process.hltParticleFlowClusterPSPPOnAA = cms.EDProducer( "PFClusterProducer",
> process.hltParticleFlowClusterECALUncorrectedPPOnAA = cms.EDProducer( "PFClusterProducer",
> ```
> 
> Should some of these be considered as well? Eg, HF-like `hltParticleFlowClusterHFForEgammaUnseeded`?

Good point. Now hltParticleFlowClusterHFForEgammaUnseeded also adopts Basic2DClusterForEachSeed.