@perrotta 2020-08-15
+1
- Special cluster builder designed for use at HLT
- In the UnitTests outputs one can observe the reduction in cpu time for the affected HLT modules
- Jenkins tests pass and show no differences in reco outputs